Move Ruby demo app into local demo
The TAS demo is going to change to use a slightly different Ruby app, so this
commit moves the old Ruby app into the local demo, which will continue to use
the same flavor.

This commit also updates the docker-compose to use a password for the pg
instance of Conjur.